Adds view tracking for feeds
This adds a new `views` module that tracks when various feeds have been viewed.  It also computes when the latest feed item was added as the `updatedAt` for a feed.

Technical note: all dates are now stored in epoch milliseconds to make serialization easier since we don't actually care about the date itself as an object.