Adds support for drag and drop in feed tree
This is pretty rough still, and needs some work to reduce duplication, etc.  That said, it allows hierarchical drag and drop without overly complicating the app.  Direct implementation of HTML5 DnD was chosen over using react-dnd because the latter library makes design choices that make it hard to cleanly determine the relative position of the dragged item vs the drop target, which is essential for a tree structure.