Adds basic import functionality
 This is a bit wasteful (2x parsing cost), however most people should see imports performed in under a second or two anyways.

A small *ahem* hack was required to prevent the popup menus from dismissing the import modal before it even displays.  This uses `onClickCapture` to shuffle the event order.  The downside is that if you click anywhere the event still bubbles up and the menu gets dismissed.  This can be solved with portals in react 16.x.