Adds undo toast instead of hidden menu option
Okay, so I don't love the name toast, but so it goes, that's what they're called.

This implements toasts the same way that modals (dialogs and menus) are currently implemented.  A notifications middleware has also been added that watches for various actions, and triggers notifications, err toasts for them.

Currently the only toast is the UndoToast for undoing the last desructive update.