Close sidebar when browserAction is triggered
In order to receive events, you need to not have a popup, and you need to trigger open/close events on sidebarAction and browserAction from a direct user event.

This means we need to jump through a few small hoops if we want an opportunity to close the sidebar.

Next we'll need a setting to let you view it as a popup or a sidebar.