Avoid initialization race conditions*
As @sb8244 mentioned the example in the documentation does not account for the case where the store has already initialized before the subscribing.

Instead of fiddling around with subscribing and unsubscribing, we now take the declarative approach and watch the store, if it's empty we render an empty screen.

Related:
https://github.com/tshaddix/webext-redux/issues/5

*: Hopefully.  The issue was most commonly visible when upgrading Firefox which isn't something that's easy to test.